\name{sos.data}
\alias{sos.data}
\docType{data}
\title{SOS pathway perturbation data}
\description{
  RT-PCR data with 9 genes in SOS pathway of  \emph{Escherichia coli} perturbed. These 9 perturbed genes are observed in the RT-PCR experiments.
}
\usage{data(sos.data)}
\format{
   \emph{sos.data} is a data frame containing 9 rows for observed genes and 9 columns for perturbations.
}
\references{
  T.S. Gardner, D.di Bernardo, D. Lorenz, and J.J. Collins. Inferring genetic networks and identifying compound mode of action via expression profiling. Science, 301(5629): 102-105, 2003.
}
\examples{
data(sos.data)
sos.data
}
\keyword{datasets}
